+++ Task2/Lab1_3/Mixin/HaveVisiblePart.h
@@ -8,10 +8,10 @@ public:
 	virtual ~IHaveVisiblePart() = default;
 
 	virtual void						SetVisible(bool value) = 0;
-	virtual void						SetVisible(unsigned index, bool value) = 0;
-	virtual bool						GetVisible(unsigned index) const = 0;
+	virtual void						SetVisible(size_t index, bool value) = 0;
+	virtual bool						GetVisible(size_t index) const = 0;
 
-	virtual void						CheckVisibleIndex(unsigned index) const = 0;
+	virtual void						CheckVisibleIndex(size_t index) const = 0;
 };
 
 
@@ -23,11 +23,11 @@ public:
 // Methods
 public:
 	void						SetVisible(bool value) override;
-	void						SetVisible(unsigned index, bool value) override;
-	bool						GetVisible(unsigned index) const override;
+	void						SetVisible(size_t index, bool value) override;
+	bool						GetVisible(size_t index) const override;
 
 protected:
-	void						CheckVisibleIndex(unsigned index) const override;
+	void						CheckVisibleIndex(size_t index) const override;
 //////////////////////////////////////////////////////////////////////
 // Data
 protected:
